pyproject.toml
[project]
name = "ai-review-equilibrium"
version = "0.1.0"
description = "Research workspace for AI reviewing equilibrium"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "pandas",
    "numpy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "tqdm",
    "python-dotenv",
    "tiktoken",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
